+++ example/go_example.dart
@@ -4,6 +4,6 @@ int twice(int a) => a * 2;
 
 main() async {
   print(await go(twice, 5));
-  Task twiceTask = asTask(twice);
+  Task twiceTask = remoteTask(twice);
   print(await twiceTask(5));
 }