+++ lib/src/go_base.dart
@@ -18,18 +18,25 @@ typedef FutureOr<R> Task<R, P>(P param);
 ///     main() async {
 ///       print(await go(twice, 5));  // => 10
 ///     }
-Future<R> go<R, P>(Task<R, P> task, P params) async {
+Future<R> go<R, P>(Task<R, P> task, P param) async {
   final receivePort = new ReceivePort();
   final isolate = await Isolate.spawn(
-      _delegate, new _DelegateParams(task, params, receivePort.sendPort));
+      _delegate, new _DelegateParams(task, param, receivePort.sendPort));
   final R result = await receivePort.first;
   await isolate.kill();
   return result;
 }
 
-R _goSync<R, P>(Task<R, P> task, P params) {
-   final Future<R> result = go<R, P>(task, params);
-   // TODO
-}
-
-Task<R, P> asTask<R, P>(Task<R, P> task) => (P params) => go(task, params);
+/// Converts [task] to a function that executes [task] on another isolate and
+/// return result.
+///
+/// Example:
+///
+///     int twice(int a) => a * 2;
+///
+///     main() async {
+///       Task twiceTask = asTask(twice);
+///       print(await twiceTask(5));  // => 10
+///       print(await twiceTask(10)); // => 20
+///     }
+Task<R, P> remoteTask<R, P>(Task<R, P> task) => (P params) => go(task, params);