+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
-## 0.0.1
+## 1.0.0
 
-- Initial version, created by Stagehand
++ `go` executes the given function in another isolate and returns results
++ `remoteTask` returns converts a task into a remote executed task